Add files via upload
Hi! Please find attached two scripts. `Tina - BoW` contains the bag of words routine and a pre-prepared template for visualisation of BoW, but we still need to work with stopwords. 

`Embeddings - Retrieved with Python` is an Rmd script where I got embeddings with Python. This scipt works much faster than in pure R. It also contains PCA on the rough embeddings I got and a graph (no phenomenal findings, I am afraid). A CSV with summed embeddings is available in the data folder.